@@ -1,5 +1,7 @@
-﻿using Microsoft.AspNetCore;
+﻿
+
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
 
 namespace AnezileAcademy
 {
@@ -7,12 +9,15 @@ namespace AnezileAcademy
     {
         public static void Main(string[] args)
         {
-            BuildWebHost(args).Run();
+            CreateDefaultBuilder(args).Run();
         }
 
-        public static IWebHost BuildWebHost(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>()
+        public static IHost CreateDefaultBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.UseStartup<Startup>();
+                })
                 .Build();
     }
 }
